fix(demo-scheduler|jv-scheduler|jv-ui-components): move the file for style config to the 'styles' folder in the jv-ui-components
