+++ packages/demo-scheduler/src/App.tsx
@@ -6,7 +6,6 @@
 
 import { useEffect, useState } from "react";
 import "@jaspersoft/jv-ui-components/dist/jv-ui.css";
-import "@jaspersoft/jv-ui-components/material-ui/JVMuiClassNameSetup";
 import {
   Authentication,
   VisualizeFactory,
