+++ packages/jv-scheduler/src/components/SchedulerMain.tsx
@@ -5,7 +5,6 @@
  */
 
 import React, { useEffect } from "react";
-import "@jaspersoft/jv-ui-components/material-ui/JVMuiClassNameSetup";
 import Tabs from "./Tabs/Tabs";
 import { useDispatch, useSelector } from "react-redux";
 import { setInitialPluginState } from "../actions/action";
